refactor and split the props in to the image directory
