+++ src/components/Image/Image.d.ts
@@ -1,0 +1,13 @@
+export interface ImageProps {
+    src: string,
+    aspectRatio?: number,
+    color?: string,
+    disableError?: boolean,
+    disableSpinner?: boolean,
+    disableTransition?: boolean,
+    errorIcon?: ReactNode,
+    imageStyle?: object,
+    loading?: ReactNode,
+    onClick?: (event:React.MouseEvent<HTMLElement> ) => any,
+    style?: object
+}
