+++ src/index.d.ts
@@ -1,16 +1,4 @@
 import { ReactNode } from "react";
-export interface ImageProps {
-    src: string,
-    aspectRatio?: number,
-    color?: string,
-    disableError?: boolean,
-    disableSpinner?: boolean,
-    disableTransition?: boolean,
-    errorIcon?: ReactNode,
-    imageStyle?: object,
-    loading?: ReactNode,
-    onClick?: (event:React.MouseEvent<HTMLElement> ) => any,
-    style?: object
-}
+import ImageProps from './components/Image/Image'
 export const ImageComponent: React.ComponentType<ImageProps>;
 export default ImageComponent;
